fix lgtm warning-v3

+++ platform/broadcom/sonic-platform-modules-dell/s5232f/sonic_platform/chassis.py
@@ -10,6 +10,7 @@
 
 try:
     import sys 
+    from __future__ import division
     from sonic_platform_base.chassis_base import ChassisBase
     from sonic_platform.sfp import Sfp
     from sonic_platform.eeprom import Eeprom
@@ -91,7 +92,7 @@ class Chassis(ChassisBase):
         change_dict['sfp'] = port_dict
         elapsed_time_ms = 0
         sleep_time_ms = 500
-        sleep_time = float(sleep_time_ms / 1000)
+        sleep_time = sleep_time_ms / 1000
 
         while True:
             for port_num in range(self.PORT_START, (self.PORT_END + 1)):
